Bring the 'new-bus' to the i386.  This extensively changes the way the i386 platform boots, it is no longer ISA-centric, and is fully dynamic. Most old drivers compile and run without modification via 'compatability shims' to enable a smoother transition.  eisa, isapnp and pccard* are not yet using the new resource manager.  Once fully converted, all drivers will be loadable, including PCI and ISA.
(Some other changes appear to have snuck in, including a port of Soren's
 ATA driver to the Alpha.  Soren, back this out if you need to.)

This is a checkpoint of work-in-progress, but is quite functional.

The bulk of the work was done over the last few years by Doug Rabson and
Garrett Wollman.

Approved by:	core

+++ sys/dev/ata/ata-disk.c
@@ -25,7 +25,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
  * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  *
- *	$Id: ata-disk.c,v 1.5 1999/03/28 18:57:18 sos Exp $
+ *	$Id: ata-disk.c,v 1.6 1999/04/10 18:53:35 sos Exp $
  */
 
 #include "ata.h"
@@ -527,11 +527,11 @@ ad_transfer(struct buf *bp)
     /* output the data */
 #if 0
     outsw(adp->controller->ioaddr + ATA_DATA,
-          (void *)((int32_t)bp->b_data + adp->donecount),
+          (void *)((uintptr_t)bp->b_data + adp->donecount),
           adp->currentsize / sizeof(int16_t));
 #else
     outsl(adp->controller->ioaddr + ATA_DATA,
-          (void *)((int32_t)bp->b_data + adp->donecount),
+          (void *)((uintptr_t)bp->b_data + adp->donecount),
           adp->currentsize / sizeof(int32_t));
 #endif
     adp->bytecount -= adp->currentsize;
@@ -590,11 +590,11 @@ oops:
         /* data ready, read in */
 #if 0
         insw(adp->controller->ioaddr + ATA_DATA,
-             (void *)((int32_t)bp->b_data + adp->donecount), 
+             (void *)((uintptr_t)bp->b_data + adp->donecount), 
 	     adp->currentsize / sizeof(int16_t));
 #else
         insl(adp->controller->ioaddr + ATA_DATA,
-             (void *)((int32_t)bp->b_data + adp->donecount), 
+             (void *)((uintptr_t)bp->b_data + adp->donecount), 
 	     adp->currentsize / sizeof(int32_t));
 #endif
         adp->bytecount -= adp->currentsize;
